requirements-build.txt
# Requirements needed for project build but not available as rosdeps
python-statemachine
keyboard
